requirement.txt
spade 
fastapi
uvicorn
python-dotenv
pydantic
aiohttp
